feat(utility): tests as docstrings
These utilities allow tests/examples
described in docstrings to be
evaluated, allowing unit-tests
to be written as docstrings.

The implementation is locked
to the more recent docstring
format used across the toolbox,
where we have an `%Example:` block.

`read_until_match` is used
for the docstring reading/block find,
while `comment_eval_wrapper`
is used as the closure for
the string evaluation.